+++ config.dist.php
@@ -41,6 +41,22 @@ const THUMBNAIL_CACHE_PATH = __DIR__ . '/data/.thumbnails';
 const DB_FILE = __DIR__ . '/data/db.sqlite';
 
 /**
+ * SQLite3 journaling mode
+ * Default: TRUNCATE (slower)
+ * Recommended: WAL (faster, but read below)
+ *
+ * If your database file is on a local disk, you will get better performance by using
+ * 'WAL' journaling instead. But it is not enabled by default as it may
+ * lead to database corruption on some network storage (eg. old NFS).
+ *
+ * @see https://www.sqlite.org/pragma.html#pragma_journal_mode
+ * @see https://www.sqlite.org/wal.html
+ * @see https://stackoverflow.com/questions/52378361/which-nfs-implementation-is-safe-for-sqlite-database-accessed-by-multiple-proces
+
+ */
+//const DB_JOURNAL_MODE = 'WAL';
+
+/**
  * WWW_URL is the complete URL of the root of this server
  *
  * If you don't define it, KaraDAV will try to auto-detects it as well as it can.
@@ -74,23 +90,6 @@ const LOG_FILE = null;
  * see https://tn123.org/mod_xsendfile/
  */
 const ENABLE_XSENDFILE = false;
-
-/**
- * Set to TRUE if you have a slow filesystem (eg. NFS/BindFS)
- *
- * This will disable directory sizes and directory last modification date
- * (all directories will appear as 0-bytes, and the modification date might
- * not be accurate). Not a huge impact, but it can appear weird to the user.
- *
- * Details: to find out the size taken by a directory, we must do the sum
- * of all files and sub-directories, which might be slow if you have lots
- * of files. Same for directory modification date, we need to find out the
- * last modification of each file in that directory.
- *
- * Note that this will not disable slow operations used for quotas, as it
- * would effectively disable quotas. You must disable each user quota.
- */
-const DISABLE_SLOW_OPERATIONS = false;
 
 /**
  * External authentication callback